Added API and updated bot.py
