+++ src/cloudlink/old_client/client.py
@@ -1,5 +1,5 @@
-from .clientRootHandlers import clientRootHandlers
 from .clientInternalHandlers import clientInternalHandlers
+from .clientRootHandlers import clientRootHandlers
 
 
 class client:
